+++ R/GF_GRAPH_OUTPUT_function.R
@@ -1,3 +1,0 @@
-GF_GRAPH_OUTPUT<-function(gs,attrname){
-
-}